Kernel: Refactor storage stack with u64 as mmap offset

+++ Kernel/Devices/BXVGADevice.h
@@ -42,7 +42,7 @@ public:
     BXVGADevice();
 
     virtual int ioctl(FileDescription&, unsigned request, FlatPtr arg) override;
-    virtual KResultOr<Region*> mmap(Process&, FileDescription&, const Range&, size_t offset, int prot, bool shared) override;
+    virtual KResultOr<Region*> mmap(Process&, FileDescription&, const Range&, u64 offset, int prot, bool shared) override;
 
     // ^Device
     virtual mode_t required_mode() const override { return 0660; }
